Cubs game varient
This is a simple-ish rework of the virus sim that functions similarly to the game that Lorna mentioned.  The true/fasle infection status has been upgrades to a evolution status, which can have the values 1, 2, 3, 4 or 5.  The player also have a new shared variable - hand - which is placed in then playerList dictionary.  This can have a value of 1, 2, or 3, which corresponds to rock, paper and scissors.  When two players with the same evoStatus meet, they read each others hands and decided whether they've won or lost.  IF they've won they got up an evoStatus level, to a maximum of 5, and if they lose they go down to a minimum of 1.  They players are given a hand on launch, and refresh their hand when they move away from their last opponent.  Not the best solution - ideally the hand would refresh after every interaction - but i had this odd issue where when two players meet and interact, one of them would sometimes refresh their hand after the other had already read it, resulting in situations where both players would think they'd won/lost.  I do not know or understand why this has happening.