- Fixed issue in player disconnect handling.
Currently, the playerList dictionary uses the players username as the key, and then plops their stats in an array as the value.  However the only information we get when a client disconnects is their websocket.  We are given this when they first connect, but previously we did not save it an where, so when a websocket disconnected we had no way of telling which player this actually corresponded to.  Logically we would use the websocket as the key in playerList, and put the username in with the values, but i am lazy and didn't want to make this big of a change.  So instead i made a second dictionary, where the websocket is used as the key and the players name as the value.  This way, when a websocket disconnects, we can use this new dictionary to identify the name of the player that it is associated with, and then use that to remove that player from playerList.

All this is to say that my last commit included the new logic for identifying the player name from the websocket and so on and so forth, but didn't actually ever fill the new dictionary with anything.  When a websocket disconnected the script tried to query an empty dictionary and would crash anyway.