Further viral strain updates
Made the infection distance of a virus a transferable trait.  Previously players would check every other place on the board to see if they were infected and within a pre-defined distance.  Now instead of referring to a pre-defined distance, they refer to a distance provided by the player being checked.  If they are then infected, they set their infection distance to the distance of the player they were infected by, which in turn can be passed on to further players.

This is all a fancy way of saying that infectious range is now being transfered from player to player in the same way that virus name is.  This effectively makes viral strains a real thing now.