Virality
Added a transferabled infection strength, and overhauled the method that the players use to decided in they've been infected or not.  It involved sigmoid graphs now.  My head hurts.

The moral of the story is that, while the possibility of being infected is still restricted to the infectious distance, a  virus with a high infectious strength have a greater chance of infected players that fall within their infection distance, and vis versa.  This opens the door to creating very infectious short range virus' and less infectious long range virus'.

More testing is needed though - i have a very funny screenshot of a player being infected despite having a 0.0% chance.  The change could have been 0.0-something%, and the player simply beat those 1 in a 10,000 odds, but i very strongly doubt that.