Improved expo.py
Initially this was a tool for my own sake, to wrap my head around who the infection curves were behaving, but now i think it might be a cool tool to give the masterclass participents so that they can ALSO wrap their head around how the infection curves are behaving. Assuming of course we don't can my admittedly over-complicated solution for that part of the sim.