Minor update to infectionLogic.py
When the player is within range of an infected player, they are now told how far away and in what direction they are.  With this, the players could now conceivably play the game text-based (I can't think of a good way to inform the player that they've run into a wall though.  I'd need to send a "hitWall = true/false" boolean from the server to the player, and i feel thats just adding too much overly-specific into into the playerList. It would be simple though - if the players position was illegal and had to be updated, set hitWall to true.  If not, set it to false.