Wall info is now shared to clients
Whenever a client first connects to the server, they are sent a one-off json file containing information on the size of the game board, and the position and size of the game walls.  This has a handful of real and potential benefits.  The main one that is currently being used is that the display client no longer needs to be pre-programmed with information on the size of the playable area and the position of walls.  This means that you can make changes to these values in decoupled_server.py, and they will be automatically sent to display.py when it connects and displayed correctly in its pygame window (with the current exception of adding or subtracting from the total number of walls).  The contents of the JSON file are also genertic enough (hopefully) to be easily read in and processed by Lucas JS client, which should make the integration between his work and mine more straight forwards.  Currently the only thing the player clients use this new information for is scaling their movement speed with respect to the size of the window, however it opens the door to a lot of interesting potential changes, such as more sophisticated roboclients that automatically avoid infected players and chase uninfected, and potentially AI implimentation.  You could also expand the text-based interface i was working on previously to actually tell you when you've run into a wall.