Started adding data visualisation.
This is something i've wanted to do since the very start of the project, but which was difficult to do within pygame (or at least i couldn't figure it out).  It looks like it'll be easier in JS.  Which is... hilarious.  The one thing in the universe thats easier to do in JS.