The Roboclient Update
The main change in this commit is a major overhaul in how the roboclients and threaded script launcher.  Instead of numerous individual sets of robotclient and infectionlogic scripts, there is now just one.  The new InfectionLogicX script generates random usernames, so now multiple roboclientX instances can be launched and connected to the same server without them interfering with each other (unless of course two of them pick the same random word out of the english dictionary as a username.  Seems unlikely).  Along with this, i have made it so the number of roboclientX instances that the launcher script launches is determined by a variable, rather than copy-pasting the "script1_thread.start()" commands over and over.  Now vwhen the launcher script is run, it will start up the server, a playable client, and however many roboclients you desire.