Spruced up the webpage.
Added some UI, that looks like a sort of crappy copy of the EIDF website.  I figured do this because the masterclass will be running on the EIDF. Could be filler, could be final, who knows.  Its scales... alright for different window sizes.  It also revamped how the game board is displayed - previously it used 3 seperate canvases that could slide around randomly when rescaling the window.  Now everything is slopped into a single canvas, which i can scale and move around very easily.  It works reasonally well.  Sometimes the player usernames clip outside of the playable area.  I pray for anyone that wants to edit it after me because the JS is a mess of "barPosX + barWidthX + margin + margin".  Its a bit vague.