Added controlClient.py
controlClient.py is a special client which, instead of providing legal player data, provides special commands which are picked up and executed by the server (you can see if you look in the script that, like the display client, it uses a special username which is picked out by the server, and then sends over the commands in place of an X coord).  These commands are as follows:-

- Cure all connected players
- Spawn a bot (At last, an actual purpose for autoClient!)
- Kill the last spawned bot
- Infect a random players with a long range, low strength virus (viruses created this way will always be called username-itus)
- Infect a random player with a short range, high strength virus
- Infect a random player with a short range, low strength virus
- Infect a random player with a long range, high strenth virus

This client would be used by a GM to wrangle some control over the game without having to ask the players to constantly disconnected/reconnect.  Now, in theory, the only reasons any player would need to disconnect and reconnect is to modify their infectionlogic.