Modified the controlClient a little
I tried in vain to squash bugs related to spawned in bots becoming detached from the console and thus becoming unkillable, but i was not totally successful.  As long as the GM doesn't wildly spam the "Spawn bots" buttons then nothing bad should happen.  Click the button, and wait for the bots to fully spawn in before clicking it again.

I had made it so that randomly infected players retain the virus name that they set.  This does come with one hangup though - When a player is infected by another,the virus name that they originally set when they joined is overwritten by their gained infection.  This means that if they are cured and then randomly infected by the GM, they retain the virus name from their previous infection.  There is almost certainly a fix for this but in the mean time, the users will just need to disconnect and reconnect.