A lot of old changes that i forgot to commit.
I don't remember what most of them actually were.  I was working on pathfinding for the autoclient but have scrapped everything i had done because it plain-and-simple wasn't working.

To get it to work, i'd been to impliment ray-marching or ray-casting or whatever its called.  Find the angle between the client and the target and check to see if the line between them intersects a wall.  If it doesn't, then great just move along tht line.  If it does, then send out an array of test-lines from the client tht end when they hit a wall.  Check to see if any of these points have a clear shot to the target. . No? Then do it again.  By my reconing it would take at most three repeats of this to find a workable path on my current map, but it is not a scalable solution.

So for example lets say that you find 20 test points each try, checking each time if any of the 20 have a direct line of sight to the target.  After three trys that 8000 test points and 8000 checks, the vast majority of which being negative.  This seems like a waste, and is partly why i didn't just rework the crappy pathfinding i have programmed before to do this instead.  Plus, 20 points is a bit on the meager side - thats 18 degree gaps between each of them.  Not good!

One alternate solution is to, of course after doing the initial "can i just see the target" check, would be to cast lines out from both the client AND the target, and then check if any of the endpoints of these lines can see each other.  that way, in the cast where three tries are needed to find a path, instead of finding 8000 test-points and doing 8000 checks, we would only need to find 40 points and do 441 checks (40 test points plus the client/target postion).  The problem that then arises is that if a path cannot be found after this first try, then the numbers instantly become ridiculous.  If each of the client and targets test points create 20 more points, then thats 800 points, and if each of those points needs to check if it has a direct line of sight to each other point, then thats ~160,000 checks.  A third try would have over a quarter of a billion checks.  Probably a bit much.

Now, a potential solution for THAT problem would be to thus; Right off the bat, before any checks are done, find some test point around the target.  Check to see if the client has a direct line of sight to either the target or any of these test points.  If not, then create test points around the client.  Can any of these points see the target/its points?  No?  Then create test points around them, and check again.  This is similar to the first method, but we've essentially given ourself a successful testpoint for free - we've identified 20 points that absolutely do have a direct line of sight to the target, and all we know need to do is find a path to one of them.

As i write this out (most as a rubber-duck test to myself), it is occuring to me that it would be far easier to just draw a picture of this.  Think of it like this - you and you friend are in a dark maze, which is so dark that you can't see your friend or the walls.  Instead you have a bow, and a infinite amount of arrows with tiny LEDs one them.  You can use these to map out the walls by shooting them off and seeing where the LEDs stop.  You're friend is also wearing an LED so once you have a clear line of sight, you'll be able to see them.  In the first method, you would first check to see if you can just see your friend right away.  If not, you would shoot out 20 arrows in all directions, and then walk to where each of them stuck in the wall.  If you still can't see your friend from any of these points you'd shoot out 20 more arrows from each point, and then walk to where each of them stuck into the wall.  If you still can't see your friend, you'd do this again.  If just so happens that in this maze you'll over ever need to do this three timess before you finally shoot an arrow that you can see your friend from, so in total you'll need to shoot 8420 arrows and do 8421 checks.  In method three (im skipping 2 because its a total non-starter at scale), your friend also has a bow with 20 arrows.  He shoots off his 20 arrows in all directions around him.  This in essensse gives you 20 ntimes the number of targets - you know that if you ever see one of these arrows, the you have a guarenteed path to your friend.  So you first check to see if you see your friend or any of this 20 arrows.  If not, you shoot off 20 of your own and walk to where they stick in the wall.  You then check to see if you can see your friend or one of their arrows.  If not, you shoot off 20 more arrows from each point.  Now we get to the benefit of this method - we said previously that in this maze, we will only ever need to repeat this first 20, move, first 20 process three times before you fire an arrow that has a direct line of side to your friend.  However, since in this setup our friend has fired 20 arrows themselves, we're already have one of these steps completed for us, so we only need to do it ourselves twice.  This means tht in total we need to fire 420 arrows and do 8841 checks - slightly more checks, but far fewer arrows.

The question then becomes whether cutting the number of arrows while not really cutting the number of checks is worth it - both take comupational time, but what takes more? If the act of firing the arrows is so cheap that reducing its numbers doesn't really do anything, then method 3 is objectively worse than method 1.