Working pathfinding for autoclient
I have finally managed to impliment the first shimmers of a half-way functional pathfinding system for the autoClient.  It works, broadly speaking, by finding some testpoints on around the player and target that have a direct line of sight to either client (so both the player and target get their own set), identifying pairs of these testpoints that have direct line of sight to each other, identifying the pair that are closest together, and then moving the autoclient in that direction.  Its not perfect, but it is able to navigate corners, and move the autoClient in a direction which puts in further from the target, but gives it a better chance of actually reaching it.  It takes funny wobbly paths but i have not seen it outright fail yet.

One issue is that it does not handle convex corners very well.  I think this is probably because i haven't figured the players radius into its pathfinding, so its trying to sneak past walls that its too big to fit around.  This is an easy fix.