Fixed noted issue where autoClient would try to pathfind through walls.
There was a false-positive coming from situation where one of the TPs was lying directly on the wall between them.  Since the wall was not technically between the two TPs, it was being passed as a legitimate path when it wasnt.

Convex corner issue still exists.  Interesting i observed an issue that i speculated would occur from the start but which i hadn't need until now - If the autoClient finds that none of its TPs can see any of the targetTPs, it can get stuck between two or more point when trying to move towards its most distance TP. I've included a screenshot of a situation wherein the autoClient got stuck trying to move towards three different TPs each.

One thing that occurs to me that might help resolve this issue, and make the pathfinding system a bit more robust, would be create more test points around the most distance test points, and see if they see the targetTPs. If they dont, then move to the second most distance TP. In the vast majority of cases it'll only need to create one extra set of TPs, but it should combat little edge bases like this.