Cleaned up path finding, still doesn't quite work.
I feel i am closer though.  I believe the main hitch now is thats im not properly considering the rotation of the upper and lower outerTP lines.
So, so we cast a mainTP line out with an angle of 0 degrees.  In this case, the upper outerTP line will obviously be above the mainTP line, and the lower will be below it.  However, if we cast a mainTP line out at 180, or anything between 90 and 180, its actually the other way around, the "upper" will be below the mainTP line and vis versa.  Since i haven't really considered this part of the puzzle carefully enough, its very possible that that is where my incorrect point translation woes are coming from.

So basically, for angles greater than 90, we need to reverse the translations in xStep1, xStep2, yStep1 and yStep2 to some degree.

After this, there is the issue that i've now entirely chopped out the bots ability to recognise when its in direct line of sight to the target player.  This obviously isn't ideal, but i hope once the first issue is sorted this should at the very least be a tractible problem.

After THAT, there is the slight issue that once the Tps are found, the check that it done to see if any of them have direct line of sight to each other does not consider player width AT ALL. In theory this isn't an issue because the player will never be guided to a point that it can't fit in, but i have a hunch this will become problematic.

Finally, once all that it well and fixed, i need to go back and actually improve the pathfinding itself.  Namely i want to add another layer of TP generator in the case where the initial round didn't find any viable paths.  This is doable, and i infact did it before, but that work got lost when i reverted to an old state a week or so ago.