logic to find the true FurthestViableTP is now fixed!
🥳

What this means is basically that the childTPs around a parentTP will now _always_ be in a position that the player can legally move to without bumping into a wall.  This is fair from finished but that one part, which has caused me such strife, is now solved.
I have also added back in the feature of the player and target "rotating" to face each other, again this time actually working.  This doesn't really have a purpose now but it will later - having this feature allows us to use magiciansStaff in magicMissile as a way of checking to see if a childTP has a line-of-sight on the target/one of the targets childTPs.